Decorate inline code in zoomed title
The zoomed page title used plain textContent and didn't render backtick
runs as mono chips like the bullets do. Extract the inline-code decoration
helpers into a shared inline-code.ts and apply the same live, caret-preserving
decoration (with IME composition handling) to ZoomedTitle.
