+++ src/components/OutlineNode.tsx
@@ -6,6 +6,7 @@ import type { Node } from "../data/schema";
 import type { TreeIndex } from "../data/tree";
 import { childrenOf } from "../data/tree";
 import { useSlashMenu } from "./slash-menu";
+import { decorate, getCaretOffset } from "./inline-code";
 
 interface OutlineNodeProps {
   node: Node;
@@ -354,86 +355,6 @@ export const OutlineNode = memo(function OutlineNode({
     </li>
   );
 });
-
-// Inline `code` runs: `like this`. Single-line, non-empty, no nested backtick.
-// Backticks are kept VISIBLE inside the chip so the source string and the
-// rendered text have identical length -- caret offsets need no source<->display
-// mapping. The stored text is always the raw markdown.
-const INLINE_CODE = /`[^`\n]+`/g;
-
-const CODE_CLASS =
-  "rounded-[4px] border border-border/60 bg-muted px-0.5 py-0.5 font-mono text-[0.85em] text-foreground";
-
-// Build the display HTML for a bullet's text. Escapes first (the text is user
-// input going into innerHTML), then wraps inline-code runs. We wrap the WHOLE
-// match -- backticks included -- so the chip's text length equals the source's;
-// that's what lets the caret-offset save/restore work without mapping source
-// positions to display positions. Backticks aren't escaped, so they survive
-// escaping and the regex still matches the already-escaped text.
-function inlineMarkupHtml(text: string): string {
-  return escapeHtml(text).replace(
-    INLINE_CODE,
-    (match) => `<code class="${CODE_CLASS}">${match}</code>`,
-  );
-}
-
-function escapeHtml(s: string): string {
-  return s.replace(/&/g, "&amp;").replace(/</g, "&lt;").replace(/>/g, "&gt;");
-}
-
-// Re-render `el` as formatted HTML, optionally keeping the caret put. The DOM
-// rebuild invalidates any (node, offset) selection, so we capture the caret as
-// an absolute character offset first and re-derive it against the new tree
-// after. Writing innerHTML only when it actually changes avoids needless
-// rebuilds (and the caret flicker they cause) when text is already current.
-function decorate(el: HTMLElement, text: string, preserveCaret: boolean): void {
-  const html = inlineMarkupHtml(text);
-  if (el.innerHTML === html) return;
-  const offset = preserveCaret ? getCaretOffset(el) : -1;
-  el.innerHTML = html;
-  if (offset >= 0) setCaretOffset(el, offset);
-}
-
-// Absolute count of characters before the collapsed caret within `el`, across
-// all text nodes (so a `<code>` chip in the middle is counted transparently).
-function getCaretOffset(el: HTMLElement): number {
-  const sel = window.getSelection();
-  if (!sel || sel.rangeCount === 0) return 0;
-  const range = sel.getRangeAt(0);
-  if (!el.contains(range.endContainer)) return 0;
-  const pre = range.cloneRange();
-  pre.selectNodeContents(el);
-  pre.setEnd(range.endContainer, range.endOffset);
-  return pre.toString().length;
-}
-
-// Drop the caret at absolute character `offset`, walking text nodes until the
-// offset falls inside one. Past-the-end lands at the very end.
-function setCaretOffset(el: HTMLElement, offset: number): void {
-  const sel = window.getSelection();
-  if (!sel) return;
-  const walker = document.createTreeWalker(el, NodeFilter.SHOW_TEXT);
-  let remaining = offset;
-  let node = walker.nextNode();
-  while (node) {
-    const len = node.textContent?.length ?? 0;
-    if (remaining <= len) {
-      const range = document.createRange();
-      range.setStart(node, remaining);
-      range.collapse(true);
-      sel.removeAllRanges();
-      sel.addRange(range);
-      return;
-    }
-    remaining -= len;
-    node = walker.nextNode();
-  }
-  const range = document.createRange();
-  range.selectNodeContents(el);
-  range.collapse(false);
-  sel.removeAllRanges();
-  sel.addRange(range);
-}
 
 // Caret at the very end / start of the bullet, measured by absolute offset so
 // the test holds whether the line is one text node or split around chips.
